requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
apscheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
bitarray==3.4.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
cryptography==45.0.3
cytoolz==1.0.1
eth-abi==5.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
exceptiongroup==1.3.0
frozenlist==1.6.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperliquid-python-sdk==0.15.0
idna==3.10
msgpack==1.1.0
multidict==6.4.4
parsimonious==0.10.0
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-core==2.33.2
python-telegram-bot==22.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
setuptools==80.9.0
sniffio==1.3.1
toolz==1.0.0
typing-extensions==4.13.2
typing-inspection==0.4.1
tzlocal==5.3.1
urllib3==2.4.0
websocket-client==1.8.0
yarl==1.20.0
